pico-library-api/requirements.txt
alembic==1.13.1
bcrypt==4.1.2
black==24.2.0
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.1.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.28
tomli==2.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
